\name{edss}
\alias{edss}
\docType{data}
\title{Data for testing ms.sev package}
\description{
Mock data for 2,000 individuals to test the ms.sev package.
}
\usage{data(edss)}
\format{
A data frame consisting of 6 columns and 2,000 rows. The rows corresponds to EDSS value, age at which the EDSS score was assessed and disease duration. 
}
\references{
To be completed. Please contact maintainer.
}
\keyword{datasets}
\keyword{ ARMSS }
\keyword{ EDSS }
\keyword{ severity}
\keyword{ MSSS }
