\name{globalArmss}
\alias{globalArmss}
\docType{data}
\title{Global ARMSS matrix}
\description{
The global ARMSS-matrix provided with the ms.sev-package
}
\usage{data(globalArmss)}
\format{
A data frame consisting of 20 columns and 30 rows, corresponding to EDSS-scores in 0.5-increments, and disease duration. The first row marks the disease duration.
}
\references{
To be completed. Please contact maintainer.
}
\keyword{datasets}
\keyword{ ARMSS }
\keyword{ EDSS }
\keyword{ severity}
\keyword{ MSSS }
