\name{globalMsss}
\alias{globalMsss}
\docType{data}
\title{Global MSSS matrix}
\description{
The global MSSS-matrix provided with the ms.sev-package
}
\usage{data(globalMsss)}
\format{
This is the new, updated MSSS-matrix. A data frame consisting of 20 columns and 30 rows, corresponding to EDSS-scores in 0.5-increments, and disease duration. The first column is corresponding to disease duration.
}
\references{
To be completed. Please contact maintainer.
}
\keyword{datasets}
\keyword{ ARMSS }
\keyword{ EDSS }
\keyword{ severity}
\keyword{ MSSS }
