\name{ms_sev-package}
\alias{ms_sev-package}
\docType{package}
\title{
Estimation of severity scores for multiple sclerosis. 
}
\description{
Estimation of severity scores for multiple sclerosis. 
}
\details{
\tabular{ll}{
Package: \tab ms_sev\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-30\cr
License: \tab GPL 2\cr
}

}
\author{
Helga Westerlind, Ali Manouchehrinia

Maintainer: Helga Westerlind <Helga.Westerlind@ki.se>

}
\keyword{ package }
