\name{oldMsss}
\alias{oldMsss}
\docType{data}
\title{the old Global MSSS matrix}
\description{
The global MSSS-matrix provided with the ms.sev-package
}
\usage{data(oldMsss)}
\format{
A data frame consisting of 20 columns and 30 rows, corresponding to EDSS-scores in 0.5-increments, and disease duration. This is the original MSSS matrix. The first column is corresponding to disease duration.
}
\references{
To be completed. Please contact maintainer.
}
\keyword{datasets}
\keyword{ ARMSS }
\keyword{ EDSS }
\keyword{ severity}
\keyword{ MSSS }
